Fixes a race condition on reconnect which caused interleaving of duplicate events.
Timeline:
1. Thread A initial connect
2. Keep-alive thread detects failure when binlog position = X
3. Keep-alive starts reconnect thread B at position X
4. Thread A does not realize that it should stop processing events
5. Thread A and B process duplicate binlogs from position X on up

Previously, the 2 ways thread A was supposed to notice the reconnect and
stop:
1. The socket's input stream gets shut down
2. It processes events only if the global connected state == true

From what I can tell, thread A was supposed to stop because the state
became disconnected, and then exit because the next read should
have yielded EOF.

This doesn't always work because:
1. The input stream is buffered so the EOF read is delayed
2. Before it reaches the end of the buffer, thread B connects
3. Thread A now processes the rest of its buffered events because thread
B updates the global connect state

The symptom of this is interleaving duplicate events. This especially
messes up the transaction logic.

This commit fixes this by giving listening threads more insight into the
connection state. It will only process events and continue reading from
the input stream if it sees that it's specific channel is open.